Create Bytes-MySQL-Backup-To-FTP-Upload.sh
Simple script to perform a mysqldump and compress backup to then upload to a FTP server. Only improvement on this would be to skip FTP all together and use sFTP or FTPS instead. Tested on a CentOS 6 machine with the FTP command issuing to a Windows Server. FTP handled with script variables credit to: http://www.stratigery.com/scripting.ftp.html. Implementing more error handling is recommended.